lcw
1、添加常用的类目统一放在CategoriesHeader.h
2、导入支付宝SDK

+++ WJXC/WJXC/Libraries/HelpTools/LTools.m
@@ -577,7 +577,7 @@
     
     if (between >= 1) { //大于一个小时需要更新
         
-        NSLog(@"需要更新融云用户信息 %@ bew:%d",oldDate,between);
+        NSLog(@"需要更新融云用户信息 %@ bew:%ld",oldDate,(long)between);
         
         return YES;
     }
